chore(fixtures) : 메뉴, 상품 fixtures 이동
