+++ src/test/java/kitchenpos/Fixtures.java
@@ -1,131 +1,66 @@
 package kitchenpos;
 
+import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.Random;
+import java.util.UUID;
 import kitchenpos.eatinorders.domain.Order;
 import kitchenpos.eatinorders.domain.OrderLineItem;
 import kitchenpos.eatinorders.domain.OrderStatus;
 import kitchenpos.eatinorders.domain.OrderTable;
 import kitchenpos.eatinorders.domain.OrderType;
-import kitchenpos.menus.domain.Menu;
-import kitchenpos.menus.domain.MenuGroup;
-import kitchenpos.menus.domain.MenuProduct;
-import kitchenpos.products.domain.Product;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.Arrays;
-import java.util.Random;
-import java.util.UUID;
 
 public class Fixtures {
-    public static final UUID INVALID_ID = new UUID(0L, 0L);
+  public static final UUID INVALID_ID = new UUID(0L, 0L);
 
-    public static Menu menu() {
-        return menu(19_000L, true, menuProduct());
-    }
+  public static Order order(final OrderStatus status, final String deliveryAddress) {
+    final Order order = new Order();
+    order.setId(UUID.randomUUID());
+    order.setType(OrderType.DELIVERY);
+    order.setStatus(status);
+    order.setOrderDateTime(LocalDateTime.of(2020, 1, 1, 12, 0));
+    order.setOrderLineItems(Arrays.asList(orderLineItem()));
+    order.setDeliveryAddress(deliveryAddress);
+    return order;
+  }
 
-    public static Menu menu(final long price, final MenuProduct... menuProducts) {
-        return menu(price, false, menuProducts);
-    }
+  public static Order order(final OrderStatus status) {
+    final Order order = new Order();
+    order.setId(UUID.randomUUID());
+    order.setType(OrderType.TAKEOUT);
+    order.setStatus(status);
+    order.setOrderDateTime(LocalDateTime.of(2020, 1, 1, 12, 0));
+    order.setOrderLineItems(Arrays.asList(orderLineItem()));
+    return order;
+  }
 
-    public static Menu menu(final long price, final boolean displayed, final MenuProduct... menuProducts) {
-        final Menu menu = new Menu();
-        menu.setId(UUID.randomUUID());
-        menu.setName("후라이드+후라이드");
-        menu.setPrice(BigDecimal.valueOf(price));
-        menu.setMenuGroup(menuGroup());
-        menu.setDisplayed(displayed);
-        menu.setMenuProducts(Arrays.asList(menuProducts));
-        return menu;
-    }
+  public static Order order(final OrderStatus status, final OrderTable orderTable) {
+    final Order order = new Order();
+    order.setId(UUID.randomUUID());
+    order.setType(OrderType.EAT_IN);
+    order.setStatus(status);
+    order.setOrderDateTime(LocalDateTime.of(2020, 1, 1, 12, 0));
+    order.setOrderLineItems(Arrays.asList(orderLineItem()));
+    order.setOrderTable(orderTable);
+    return order;
+  }
 
-    public static MenuGroup menuGroup() {
-        return menuGroup("두마리메뉴");
-    }
+  public static OrderLineItem orderLineItem() {
+    final OrderLineItem orderLineItem = new OrderLineItem();
+    orderLineItem.setSeq(new Random().nextLong());
+    return orderLineItem;
+  }
 
-    public static MenuGroup menuGroup(final String name) {
-        final MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setId(UUID.randomUUID());
-        menuGroup.setName(name);
-        return menuGroup;
-    }
+  public static OrderTable orderTable() {
+    return orderTable(false, 0);
+  }
 
-    public static MenuProduct menuProduct() {
-        final MenuProduct menuProduct = new MenuProduct();
-        menuProduct.setSeq(new Random().nextLong());
-        menuProduct.setProduct(product());
-        menuProduct.setQuantity(2L);
-        return menuProduct;
-    }
-
-    public static MenuProduct menuProduct(final Product product, final long quantity) {
-        final MenuProduct menuProduct = new MenuProduct();
-        menuProduct.setSeq(new Random().nextLong());
-        menuProduct.setProduct(product);
-        menuProduct.setQuantity(quantity);
-        return menuProduct;
-    }
-
-    public static Order order(final OrderStatus status, final String deliveryAddress) {
-        final Order order = new Order();
-        order.setId(UUID.randomUUID());
-        order.setType(OrderType.DELIVERY);
-        order.setStatus(status);
-        order.setOrderDateTime(LocalDateTime.of(2020, 1, 1, 12, 0));
-        order.setOrderLineItems(Arrays.asList(orderLineItem()));
-        order.setDeliveryAddress(deliveryAddress);
-        return order;
-    }
-
-    public static Order order(final OrderStatus status) {
-        final Order order = new Order();
-        order.setId(UUID.randomUUID());
-        order.setType(OrderType.TAKEOUT);
-        order.setStatus(status);
-        order.setOrderDateTime(LocalDateTime.of(2020, 1, 1, 12, 0));
-        order.setOrderLineItems(Arrays.asList(orderLineItem()));
-        return order;
-    }
-
-    public static Order order(final OrderStatus status, final OrderTable orderTable) {
-        final Order order = new Order();
-        order.setId(UUID.randomUUID());
-        order.setType(OrderType.EAT_IN);
-        order.setStatus(status);
-        order.setOrderDateTime(LocalDateTime.of(2020, 1, 1, 12, 0));
-        order.setOrderLineItems(Arrays.asList(orderLineItem()));
-        order.setOrderTable(orderTable);
-        return order;
-    }
-
-    public static OrderLineItem orderLineItem() {
-        final OrderLineItem orderLineItem = new OrderLineItem();
-        orderLineItem.setSeq(new Random().nextLong());
-        orderLineItem.setMenu(menu());
-        return orderLineItem;
-    }
-
-    public static OrderTable orderTable() {
-        return orderTable(false, 0);
-    }
-
-    public static OrderTable orderTable(final boolean occupied, final int numberOfGuests) {
-        final OrderTable orderTable = new OrderTable();
-        orderTable.setId(UUID.randomUUID());
-        orderTable.setName("1번");
-        orderTable.setNumberOfGuests(numberOfGuests);
-        orderTable.setOccupied(occupied);
-        return orderTable;
-    }
-
-    public static Product product() {
-        return product("후라이드", 16_000L);
-    }
-
-    public static Product product(final String name, final long price) {
-        final Product product = new Product();
-        product.setId(UUID.randomUUID());
-        product.setName(name);
-        product.setPrice(BigDecimal.valueOf(price));
-        return product;
-    }
+  public static OrderTable orderTable(final boolean occupied, final int numberOfGuests) {
+    final OrderTable orderTable = new OrderTable();
+    orderTable.setId(UUID.randomUUID());
+    orderTable.setName("1번");
+    orderTable.setNumberOfGuests(numberOfGuests);
+    orderTable.setOccupied(occupied);
+    return orderTable;
+  }
 }
